feat: migrate to ci v1
