--- config.yaml
@@ -1,60 +1,0 @@
-# Copyright 2022 Canonical Ltd.
-# See LICENSE file for licensing details.
-options:
-  cpu:
-    description: |
-      K8s cpu resource limit, e.g. "1" or "500m". Default is unset (no limit). This value is used
-      for the "limits" portion of the resource requirements (the "requests" portion is
-      automatically deduced from it).
-      See https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/
-    type: string
-  memory:
-    description: |
-      K8s memory resource limit, e.g. "1Gi". Default is unset (no limit). This value is used
-      for the "limits" portion of the resource requirements (the "requests" portion is
-      automatically deduced from it).
-      See https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/
-    type: string
-  ingestion-rate-mb:
-    description: |
-      Per-user ingestion rate limit (MB/s).
-      This config option matches exactly Loki's `ingestion_rate_mb`, except that it is an integer here
-      (Loki takes a float).
-      This same value is used internally for setting `per_stream_rate_limit`. Loki uses a default of 3 for
-      `ingestion_rate_mb`, but 4 for `per_stream_rate_limit`. For this reason we use 4 as the default here.
-      
-      Ref: https://grafana.com/docs/loki/latest/configure/#limits_config
-    type: int
-    default: 4
-  ingestion-burst-size-mb:
-    description: |
-      This config option matches exactly Loki's `ingestion_burst_size_mb`, except that it is an integer here
-      (Loki takes a float).
-      This same value is used internally for setting `per_stream_rate_limit_burst`. Loki uses a default of 6 for 
-      `ingestion_burst_size_mb`, but 15 for `per_stream_rate_limit_burst`. For this reason we use 15 as the default
-      here.
-      
-      Ref: https://grafana.com/docs/loki/latest/configure/#limits_config
-    type: int
-    default: 15
-  retention-period:
-    description: |
-      Sets a global retention period, in days, for log streams in Loki.
-      The minimum retention period is 1 day, and a value of 0 (default) means "infinity" (disables retention).
-      Loki will not be cleaning up logs if duration is set to 0.
-      This config maps directly to the loki `compactor.retention_enabled` configuration option, which is set to `false` when no retention period is defined.
-      Specifying retention periods for individual streams is not currently supported.
-
-      Ref: https://grafana.com/docs/loki/latest/operations/storage/retention/
-    type: int
-    default: 0
-  reporting-enabled:
-    description: |
-      When disabled, Loki will be configured to not send anonymous usage statistics to stats.grafana.org.
-      It is very helpful to the Grafana project, so please leave this enabled.
-      
-      When enabled, Loki will use its default values for analytics.
-      
-      Ref: https://grafana.com/docs/loki/latest/configure/#analytics
-    type: boolean
-    default: true
