--- doc/adr/README.md
@@ -1,37 +1,0 @@
-ADR template is based on https://adr.github.io/madr/examples.html.
-
-```markdown
-# ADR topic
-
-## Context and Problem Statement
-
-
-## Considered Options
-
-- Option 1
-- Option 2
-
-## Decision Outcome
-
-Chosen option: ..., because ...
-
-### Consequences
-
-* Good, because ...
-* Good, because ...
-* Bad, because ...
-
-## Pros and Cons of the Options
-
-### Option 1
-
-* Good, because ...
-* Bad, because ...
-
-### Option 2
-
-* Good, because ...
-* Bad, because ...
-
-## More Information
-```
