--- metadata.yaml
@@ -1,107 +1,0 @@
-# Copyright 2021 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-# For a complete list of supported options, see:
-# https://discourse.charmhub.io/t/charm-metadata-v2/3674/15
-
-name: loki-k8s
-
-assumes:
-  - k8s-api
-
-  # Juju 3.0.3+ needed for secrets and open-port
-  - juju >= 3.0.3
-
-summary: |
-  Loki is a set of components that can be composed into a fully featured logging stack.
-
-description: |
-  Loki for Kubernetes cluster
-
-maintainers:
-    - Jose Massón <jose.masson@canonical.com>
-
-website: https://charmhub.io/loki-k8s
-source: https://github.com/canonical/loki-k8s-operator
-issues: https://github.com/canonical/loki-k8s-operator/issues
-docs: https://discourse.charmhub.io/t/loki-k8s-docs-index/5228
-
-containers:
-  loki:
-    resource: loki-image
-    mounts:
-      - storage: active-index-directory
-        location: /loki/boltdb-shipper-active
-      - storage: loki-chunks
-        location: /loki/chunks
-  node-exporter:
-    resource: node-exporter-image
-    mounts:
-      - storage: active-index-directory
-        location: /loki/boltdb-shipper-active
-      - storage: loki-chunks
-        location: /loki/chunks
-
-# We do not need separate storages. TODO: In the next breaking change for Loki,
-# switch to having just one persisted storage, e.g. `/loki/persisted`.
-storage:
-  active-index-directory:
-    type: filesystem
-    description: Mount point in which Loki will store index
-  loki-chunks:
-    type: filesystem
-    description: Mount point in which Loki will store chunks (objects)
-
-provides:
-  logging:
-    interface: loki_push_api
-  grafana-source:
-    interface: grafana_datasource
-    optional: true
-  metrics-endpoint:
-    interface: prometheus_scrape
-  grafana-dashboard:
-    interface: grafana_dashboard
-  send-datasource:
-    interface: grafana_datasource_exchange
-    description: |
-      Integration to share with other COS components this charm's grafana datasources, and receive theirs.
-
-requires:
-  alertmanager:
-    interface: alertmanager_dispatch
-  ingress:
-    interface: ingress_per_unit
-    limit: 1 # Since this is the ingress there is no point in having more than one.
-  certificates:
-    interface: tls-certificates
-    limit: 1
-    description: |
-      Certificate and key files for the loki server.
-  catalogue:
-    interface: catalogue
-  charm-tracing:
-    description: |
-      Enables sending charm traces to a distributed tracing backend such as Tempo.
-    limit: 1
-    interface: tracing
-  workload-tracing:
-    description: |
-      Enables sending workload traces to a distributed tracing backend such as Tempo.
-    limit: 1
-    interface: tracing
-
-peers:
-  replicas:
-    interface: loki_replica
-
-resources:
-  loki-image:
-    type: oci-image
-    description: Loki OCI image
-    #upstream-source: ghcr.io/canonical/loki:2.7.4
-    upstream-source: docker.io/ubuntu/loki:2-22.04
-  node-exporter-image:
-    type: oci-image
-    description: Node-exporter OCI image
-    upstream-source: docker.io/prom/node-exporter:v1.7.0
