--- requirements.txt
@@ -1,26 +1,0 @@
-cosl>=0.0.47
-# pinned to 2.16 as 2.17 breaks our unittests
-ops
-kubernetes
-requests
-lightkube
-lightkube-models
-aiohttp
-jsonschema
-lightkube
-lightkube-models
-
-# Cryptography
-# Deps: tls_certificates
-cryptography
-
-# deps: tracing, charm_tracing
-pydantic>=2.9.0
-importlib-metadata==6.0.0
-
-opentelemetry-api==1.21.0
-opentelemetry-exporter-otlp-proto-common==1.21.0
-opentelemetry-exporter-otlp-proto-http==1.21.0
-opentelemetry-proto==1.21.0
-opentelemetry-sdk==1.21.0
-opentelemetry-semantic-conventions==0.42b0
