+++ tests/resources/alert.rule
