+++ tests/unit/conftest.py
@@ -1,6 +1,16 @@
-from unittest.mock import patch
+from unittest.mock import PropertyMock, patch
 
+import ops
 import pytest
+from charms.tempo_coordinator_k8s.v0.charm_tracing import charm_tracing_disabled
+from ops.testing import Context
+from scenario import Container, Exec
+
+from charm import LokiOperatorCharm
+
+
+def tautology(*_, **__) -> bool:
+    return True
 
 
 @pytest.fixture(autouse=True)
@@ -10,3 +20,33 @@ def patch_buffer_file_for_charm_tracing(tmp_path):
         str(tmp_path / "foo.json"),
     ):
         yield
+
+
+@pytest.fixture
+def loki_charm(tmp_path):
+    with patch.multiple(
+        "charm.KubernetesComputeResourcesPatch",
+        _namespace=PropertyMock("test-namespace"),
+        _patch=PropertyMock(tautology),
+        is_ready=PropertyMock(tautology),
+    ):
+        with patch("socket.getfqdn", new=lambda *args: "fqdn"):
+            with patch("lightkube.core.client.GenericSyncClient"):
+                with charm_tracing_disabled():
+                    yield LokiOperatorCharm
+
+
+@pytest.fixture
+def context(loki_charm):
+    return Context(loki_charm)
+
+
+@pytest.fixture(scope="function")
+def loki_container():
+    return Container(
+        "loki",
+        can_connect=True,
+        execs={Exec(["update-ca-certificates", "--fresh"], return_code=0)},
+        layers={"loki": ops.pebble.Layer({"services": {"loki": {}}})},
+        service_statuses={"loki": ops.pebble.ServiceStatus.INACTIVE},
+    )
