+++ tox.ini
@@ -3,15 +3,17 @@
 [tox]
 skipsdist=True
 skip_missing_interpreters = True
-envlist = lint, static-{charm,lib}, unit, scenario
+envlist = lint, static, unit
 
 [vars]
 src_path = {toxinidir}/src
 tst_path = {toxinidir}/tests
 lib_path = {toxinidir}/lib/charms/loki_k8s
 all_path = {[vars]src_path} {[vars]tst_path} {[vars]lib_path}
+uv_flags = --frozen --isolated --extra=dev
 
 [testenv]
+allowlist_externals = uv
 basepython = python3
 setenv =
   PYTHONPATH = {toxinidir}:{toxinidir}/lib:{[vars]src_path}
@@ -19,92 +21,49 @@ setenv =
   PY_COLORS=1
 passenv =
   PYTHONPATH
-  HOME
-  PATH
-  CHARM_BUILD_DIR
-  MODEL_SETTINGS
-  HTTP_PROXY
-  HTTPS_PROXY
-  NO_PROXY
+  CHARM_PATH
+
+[testenv:lock]
+description = Update uv.lock with the latest deps
+commands =
+  uv lock --upgrade --no-cache
+
+[testenv:lint]
+description = Lint the code
+commands =
+    uv run {[vars]uv_flags} ruff check {[vars]all_path}
+
+[testenv:static]
+description = Run static checks
+allowlist_externals =
+  {[testenv]allowlist_externals}
+  /usr/bin/env
+commands =
+    uv run {[vars]uv_flags} pyright {[vars]all_path}
+    /usr/bin/env sh -c 'for m in $(git diff main --name-only {[vars]lib_path}); do if ! git diff main $m | grep -q "+LIBPATCH\|+LIBAPI"; then echo "You forgot to bump the version on $m!"; exit 1; fi; done'
 
 [testenv:fmt]
-description = Apply coding style standards to code
-deps =
-    black
-    ruff
+description = "Format the code"
 commands =
-    ruff check --fix {[vars]all_path}
-    black {[vars]all_path}
-
-# codespell pinned cause version 2.3.0 mistakenly considers joined words such as "assertIn" invalid
-[testenv:lint]
-description = Check code against coding style standards
-deps =
-    black
-    ruff
-    codespell<2.3.0
-commands =
-    codespell {[vars]lib_path}
-    codespell . --skip .git --skip .tox --skip build --skip lib --skip venv --skip .mypy_cache
-    ruff check {[vars]all_path}
-    black --check --diff {[vars]all_path}
-
-[testenv:static-{charm,lib}]
-description = Run static analysis checks
-deps =
-    cosl
-    pyright
-    charm: -r{toxinidir}/requirements.txt
-    lib: git+https://github.com/canonical/operator#egg=ops
-commands =
-    charm: pyright {[vars]src_path} {posargs}
-    lib: pyright --pythonversion 3.8 {[vars]lib_path} {posargs}
-    lib: /usr/bin/env sh -c 'for m in $(git diff main --name-only {[vars]lib_path}); do if ! git diff main $m | grep -q "+LIBPATCH\|+LIBAPI"; then echo "You forgot to bump the version on $m!"; exit 1; fi; done'
-allowlist_externals = /usr/bin/env
+    uv run {[vars]uv_flags} ruff check --fix-only {[vars]all_path}
 
 [testenv:unit]
 description = Run unit tests
-allowlist_externals=/usr/bin/env
-deps =
-    pytest
-    coverage[toml]
-    responses
-    fs
-    -r{toxinidir}/requirements.txt
+allowlist_externals=
+    {[testenv]allowlist_externals}
+    /usr/bin/env
 commands =
-    /usr/bin/env sh -c 'stat cos-tool-amd64 > /dev/null 2>&1 || curl -L -O https://github.com/canonical/cos-tool/releases/latest/download/cos-tool-amd64'
-    /usr/bin/env sh -c 'chmod 755 cos-tool-amd64'
-    coverage run \
-      --source={[vars]src_path},{[vars]lib_path} \
-      -m pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tst_path}/unit
-    coverage report
-
-[testenv:scenario]
-description = Scenario tests
-deps =
-    pytest
-    ops[testing]
-    -r{toxinidir}/requirements.txt
-commands =
-    pytest -v --tb native {[vars]tst_path}/scenario --log-cli-level=INFO -s {posargs}
-
-[testenv:integration]
-description = Run integration tests
-deps =
-    aiohttp
-    juju
-    pytest
-    pytest-operator
-    cosl
-    minio
-commands =
-    pytest -v --tb native --log-cli-level=INFO --color=yes -s {posargs} {toxinidir}/tests/integration
+    /usr/bin/env sh -c 'stat cos-tool-amd64 > /dev/null 2>&1 || curl -L -O https://github.com/canonical/cos-tool/releases/latest/download/cos-tool-amd64 && chmod 755 cos-tool-amd64'
+    uv run {[vars]uv_flags} coverage run --source={[vars]src_path},{[vars]lib_path} -m pytest \
+        {[vars]tst_path}/unit {posargs}
+    uv run {[vars]uv_flags} coverage report
 
 [testenv:interface]
 description = Run interface tests
-deps =
-    pytest
-    -r{toxinidir}/requirements.txt
-    pytest-interface-tester
 commands =
-    pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tst_path}/interface
+    uv run {[vars]uv_flags} pytest --exitfirst {[vars]tst_path}/interface {posargs}
+
+[testenv:integration]
+description = Run integration tests
+commands =
+    uv run {[vars]uv_flags} pytest --exitfirst {[vars]tst_path}/integration {posargs}
